+++ libindy_vdr/src/resolver/mod.rs
@@ -2,6 +2,7 @@ pub mod resolver;
 
 pub mod did;
 pub mod did_document;
+pub mod types;
 pub mod utils;
 
 pub use self::resolver::{PoolResolver, PoolRunnerResolver};
