+++ libindy_vdr/src/resolver/resolver.rs
@@ -1,66 +1,11 @@
-use serde_json::Value as SJsonValue;
-
 use super::did::DidUrl;
-use super::did_document::DidDocument;
+use super::types::*;
 use super::utils::*;
 
 use crate::common::error::prelude::*;
 
 use crate::ledger::RequestBuilder;
 use crate::pool::{Pool, PoolRunner, RequestResult, TimingResult};
-
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
-#[serde(rename_all = "camelCase")]
-pub enum Result {
-    DidDocument(DidDocument),
-    Content(SJsonValue),
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
-#[serde(rename_all = "camelCase")]
-pub enum Metadata {
-    DidDocumentMetadata(DidDocumentMetadata),
-    ContentMetadata(ContentMetadata),
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ContentMetadata {
-    pub node_response: SJsonValue,
-    pub object_type: String,
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct DidDocumentMetadata {
-    pub node_response: SJsonValue,
-    pub object_type: String,
-    pub self_certification_version: Option<i32>,
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct ResolutionResult {
-    pub did_resolution_metadata: Option<String>,
-    pub did_document: Option<SJsonValue>,
-    pub did_document_metadata: Option<DidDocumentMetadata>,
-}
-
-#[derive(Serialize, Deserialize, PartialEq, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct DereferencingResult {
-    pub dereferencing_metadata: Option<String>,
-    pub content_stream: Option<SJsonValue>,
-    pub content_metadata: Option<ContentMetadata>,
-}
-
-pub trait Resolver {
-    fn resolve(&self, did_url: &str) -> VdrResult<String>;
-
-    fn dereference(&self, did_url: &str) -> VdrResult<String>;
-}
-
-pub trait CallbackResolver {}
 
 /// DID (URL) Resolver interface for a pool compliant with did:indy method spec
 pub struct PoolResolver<T: Pool> {
